Deduplicate subprocess echo test helper
