+++ test/common/subprocess/test_run_command_live.py
@@ -11,6 +11,7 @@ from time import monotonic
 import pytest
 
 from scinoephile.common.subprocess import run_command_live
+from scinoephile.common.testing import _echo_command
 
 
 def test_run_command_live_success_list():
@@ -166,23 +167,3 @@ def test_run_command_live_non_utf8_output():
     assert exitcode == 0
     assert stdout == "ÿ"
     assert stderr == ""
-
-
-def _echo_command(*arguments: str) -> list[str]:
-    """Build a portable command that echoes arguments without shell expansion.
-
-    Arguments:
-        *arguments: arguments to echo
-    Returns:
-        command argument list
-    """
-    return [
-        sys.executable,
-        "-c",
-        (
-            "import sys; "
-            "sys.stdout.reconfigure(encoding='utf-8'); "
-            "print(' '.join(sys.argv[1:]))"
-        ),
-        *arguments,
-    ]
